Add `allowFunctionEvaluation` prop.
The logic here is a bit heavy. In summary, we need to keep track of whether we generated the `onSubmitValueParser` callback. If we did, then we can respond to changes to `allowFunctionEvaluation`, otherwise we ignore it.

This process moved onSubmitValueParser into `JsonTree`'s state, because we need to be able to modify and store it.